% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEploidR.R
\name{haplotypePainter}
\alias{haplotypePainter}
\title{Painting haplotype according the reference panel}
\usage{
haplotypePainter(
  posteriorProbabilities,
  title = "",
  labelScaling,
  numberOfInbreeding = 0
)
}
\arguments{
\item{posteriorProbabilities}{Posterior probabilities matrix with the size of
number of loci by the number of reference strain.}

\item{title}{Figure title.}

\item{labelScaling}{Scaling parameter for plotting.}

\item{numberOfInbreeding}{Number of inbreeding strains copying from.}
}
\description{
Plot the posterior probabilities of a haplotype given the
 refernece panel.
}
